cherry pick 59b234b from https://gitee.com/neusoft_group_co_ltd_cc/FixSetDateTaken/pulls/5641 OpenHarmony-5.0.1-Release分支的MTP与PTP功能实现

+++ frameworks/services/media_mtp/include/storage.h
@@ -37,7 +37,7 @@ public:
     void SetFilesystemType(uint16_t fileSytemType);
     void SetAccessCapability(uint16_t accessCapability);
     void SetMaxCapacity(uint64_t maxCapacity);
-    void SetFreeSpaceInBytes(uint32_t freeSpaceInBytes);
+    void SetFreeSpaceInBytes(uint64_t freeSpaceInBytes);
     void SetFreeSpaceInObjects(uint32_t freeSpaceInObjects);
     void SetStorageDescription(std::string storageDescription);
     void SetVolumeIdentifier(std::string volumeIdentifier);
